requirements.txt
APScheduler
beautifulsoup4
Jinja2
lxml
pydantic
python-dotenv
PyYAML
requests
pytest
